bash_aliases
Made bashrc chain off bash_profile
Added bash_aliases
Updated createLinks to better support new files for dotfile system

+++ createLinks.sh
@@ -1,12 +1,24 @@
 #!/bin/sh
 
 function linkdot() {
-	if [ -f ~/.$1 ]; then
+	# if it's not already a symlink, but does exist
+	if ! [ -L ~/.$1 ] && [ -f ~/.$1 ]; then
+		# a new file for dotfiles?
+		if ! [ -f ~/.dotfiles/$1 ]; then
+			echo "Copying new $1 to .dotfiles"
+			cp ~/.$1 ~/.dotfiles/$1
+		fi
+
+		echo "Backing up old $1 to $1.bak"
 		mv ~/.$1 ~/.$1.bak
 	fi
-	ln -s ~/.dotfiles/$1 ~/.$1
+
+	if ! [ -L ~/.$1 ]; then
+		ln -s ~/.dotfiles/$1 ~/.$1
+	fi
 }
 
+linkdot bash_aliases
 linkdot bash_profile
 linkdot bashrc
 linkdot gitconfig
